moved logging testing into its own file
